Create base ChangeQueue interface
