+++ undofx/src/main/java/org/fxmisc/undo/impl/LinearChangeQueue.java
@@ -1,23 +1,7 @@
 package org.fxmisc.undo.impl;
 
-public interface LinearChangeQueue<C> {
-
-    interface QueuePosition {
-        boolean isValid();
-    }
-
-    boolean hasNext();
-
-    boolean hasPrev();
-
-    C next();
-
-    C prev();
+public interface LinearChangeQueue<C> extends ChangeQueue<C> {
 
     @SuppressWarnings({"unchecked"})
     void push(C... changes);
-
-    QueuePosition getCurrentPosition();
-
-    void forgetHistory();
 }
